requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
langgraph==0.0.52
requests==2.32.3
alembic==1.14.1
psycopg[binary]==3.2.13
